requirements/local.txt
-r base.txt
-e .
-e ../base
-e ../login
beautifulsoup4
gitpython
hgapi
ipdb
psycopg2
pytest-django
pytest-flakes
pytest-pep8
pyyaml
walkdir
